client_live_voice/requirements.txt
aiohttp==3.11.18
azure-core==1.34.0
azure-cosmos==4.7.0
azure-identity==1.22.0
certifi==2025.4.26
cffi==1.17.1
cryptography==44.0.3
flask==3.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==2.2.5
pycparser==2.22
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
sounddevice==0.5.1
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0